add optional "shouldDecode" argument to Reader.gets()
By default, if a Reader is configured with an encoding, all string
replies provided by reader.gets() will be decoded with the configured
encoding. There are some replies that we never want to decode.
This allows the client/user to override the default behavior
in these specific situations without having to manage multiple
reader instances.