Add ability to control how unicode decoding errors are handled
Adds an errors argument to Reader objects to be used in conjunction with the
encoding argument. Valid values for errors are 'strict' (the default),
'replace', 'ignore', or 'backslashreplace'. These four values are defined
by Python: https://docs.python.org/3/howto/unicode.html#the-string-type

Note this is a backwards incompatible change. Prior to this change
hiredis-py would silence UnicodeDecodeErrors and return the original bytes
object. After this change, the default behavior is to raise the
UnicodeDecodeError when a decoding error is encountered. This matches
Python's default behavior. The user can alter the default behavior by
specifying a different value to errors just like normal Python.