Return Redis sets as Python lists
In some (rare) cases, the sets from a Redis response contain nested
types that are not hashable in Python, for example maps. To handle these
cases uniformly, always return Python lists for Redis sets. The elements
will still be unique, relying on the correctness of data arriving from
the server.